requirements.txt
# Pin exact versions for reproducible builds
pystray==0.19.5
Pillow==10.4.0
pyinstaller==6.11.1
